V2.0 (#13)
* Added spinner hold to videos setup

* Updated spinner and readme file

* Rework of setup logic for easier scaling in future.

* Motified run script order with new names and changes

* Added correct hotspot script to run order.

* Added echo feedback to start of kisok mode setup.

* Add reboot message to end of main setup script.

* Fixed videos script. Install comanion question was writing instead of video setup question

* Companion updater now on latest stable

* Updated service seed configs for images and videos to target DISPLAY 0 for new x11 workflow.

* Added configs for x11 service

* Added x11 setup script

* Set new scripts to script run order. updated x11 session conf - change WantedBy to default. Added start script to ask to setup display.

* Changed to a systemside systemd service

* removed files from old user systemd service.

* workflow completed for x11 dual screen. HDMI0 is image and video playback screen. HDMI1 is timer screen. When only one screen in connected, it reverts back to single screen operation with the timer allways on and under.

* removed the usbc console setup script from main setup run uder until fix. currently hangs on shutdown.

* Added hotspot script into main set. Removed usbc-console script from main but left in scripts area for access. Removed old LABWC and hide cursor scripts.

* Removed duplicate hotspot script in script run order conf.

* Removed old scripts from labwc v1.0 and updates scripts

* Added fix for known rpi5 gpu issue.

* Added local time mount point in docker compose file for time zone portabiity.

* Reverted TZ as Docker needs that envoiment varible even with vol localtime

* Added companion update so user can select version

* Adding spinner to clean out put to longest tasks

* Adding spinnner to disaply and raspotify install

* Added apt autoremoved for clean up

* Added spinner animation to clean up whole install process.

* Test push for VSC.

* Testing cleaning up formatting in main setup script.

* Cleaning up formatting

* Cleaned up formatting for xinintrc

* Anabled greetd

* Cleaning up display script

* Corrected format

* Added chromium triger-kill-trigger sequence to .xinit to forve fullscreen kiosk on second display

* Reverted back to non kiosk mode

+++ scripts/companion-setup.sh
@@ -1,17 +1,33 @@
 #!/usr/bin/env bash
 
-#Get install script install.sh from github.
-curl -O https://raw.githubusercontent.com/bitfocus/companion-pi/main/install.sh
+# Gets spinner funtion
+# spinner example.
+#       Command > /tmp/log.txt 2>&1 &
+#       spinner $! "Doing something..." /tmp/log.txt
+# shellcheck disable=SC1091
+source $HOME/git-keys/spinner.sh
 
-#Make the script executable
-chmod +x install.sh
+# Asks to run script
+echo -e -n "\n\033[1m"Install Companion?"\033[0m (y/n): "
+read -r -p "" choice </dev/tty
+if [[ "$choice" =~ ^[Yy]$ ]]; then
 
-#Sets stable build and runs the script with root privileges using bash
-sudo COMPANION_BUILD=stable bash install.sh
-sudo companion-update
+	# Get install script install.sh from github.
+	curl -O https://raw.githubusercontent.com/bitfocus/companion-pi/main/install.sh
 
-#Removes the script from show-pi directory
-rm install.sh
+  # Make the script executable
+	chmod +x install.sh
 
-# Starts the compannion service
-sudo systemctl start companion.service
+	# Sets stable build and runs the script with root privileges using bash
+	sudo COMPANION_BUILD=stable bash install.sh >/tmp/log.txt 2>&1 &
+	spinner $! "Installing Companion. Be ready to select build..." /tmp/log.txt
+	sudo companion-update
+
+	# Removes the script from show-pi directory
+	rm install.sh
+
+	# Starts the compannion service
+	sudo systemctl start companion.service
+else
+	echo -e "Skipped Companion install\n"
+fi
